request #10614: Inconsistencies in database breaks REST api
Create a user group. In a tracker, bind a user selectbox (assigned to)
to this user group. Now remove this user group. In the api explorer: GET
/trackers/:id => a notice prevents you to get results.

This happens only on dev instances since prod should not display errors.

The best way would be to clean the database whenever we delete a user
group. However, the way the table is designed does not allow to do it
nicely (if only they were designed with 3NF… #sadpanda). Therefore a
good enough solution is to ignore bad entries.

Change-Id: Ia869bf28762d75bc9b00145eca1375d013bc70c9