Display results
This is a part of story #10495 - have base cross tracker search widget
(step 1)

In order to ease review for now the results are only loaded at first
page display.
Users are not clickable yet, and displayed date is not formatted.
Will be done in dedicated commit.

If you click on save report or if you update your tracker list,
results won't refresh, these points will be done in dedicated commit

Change-Id: I165a407fa650dfca088facabd69fff0aa5199b84