request #10615: Add tooltip back to dashboards
Since new dashboards, the tooltip on cross references are not anymore
displayed.

To fix this ASAP, the proposal is to bring back the old mechanism as is.
This has major drawbacks:
* Tooltip look 'n feel is ugly.
* Loading of tooltips in recent ES7 code is dirty

However this allows to remove a regression and decrease the frustration
of users in a minimal amount of time, instead of investigating about a
rewrite of our Tooltip library (get rid of jQuery, use dedicated
foundation like popper.js, …).

Change-Id: Ibc25e0001a80733c9086c5030e7cc1bc0684470a