Translate placeholder "Add labels"
This is part of story #10498: Add labels on pull requests

As choosing the right library to do our i18n stuff in javascript is out of the
scope of this story, I suggest that we keep the existing mechanism
codendi.getText. This will be easily replaced after we choose the right
library, if any.

Change-Id: Idf9229c720913ec90ba0edfe0e3297f0b3f9ea30