request #10625: Use Chrome/Chromium headless instead of PhantomJS
The binary of Chrome/Chromium can have a different name dependending of the OS/
distribution used making it not that easy to properly detect if Chrome or Chromium
is available in the environnement.
Since it is generally easier to install Chromium on Linux since it often is in
the distribution repositories, this contribution consider than when you run a
Linux you use Chromium else you use Chrome.

In any case with this contribution you need to have a recent version (aka a supported
one) of Chrome/Chromium to launch 'npm run test'.

Tests for the tracker artifact modal are now run independently from kanban.

Change-Id: Ic4757648d9414f2e6b462dd500381bb404f181ae