Display tooltip in cross tracker widget
This is a part of story #10495 - have base cross tracker search widget

When you mouseover an artfiact in cross tracker widget, you should see a
tooltip (if tooltip is configured in the given tracker).

Furthermore, we rethrow error so that it pops in the console else it is
pretty difficult to know the reason why the widget does not work (typo
in source code, failing request, …).

Change-Id: I44d99eae93709436f6739f756f708bd1252b60e7