bugs #40346 The trackers must not be imported at service activation
How to test:
As a project admin:
  - Go in a project without TTM enabled
  - Enable it from the admin section of your project
  - Go in the service
--> A message explains you that the plugin is not fully configured
and invites you, either to go in the admin section to set your trackers
manually or to let Tuleap do it for you.

Now as a non-admin project member:
  - Go in the service
--> A message explains you that the plugin is not fully configured
but you have to come back later.

Re-connect as a project admin
  - Go in the admin section
--> you can see that your tracker ids are not set up.

  - Go back to the previous message and click "start TestManagement"
--> Tuleap set up the needed trackers and, once it's done, redirects you
to the campaign view.
--> A message appears saying that the plugin has been successfully configured.

Change-Id: If97a34ca7ea659df3ce0c27850f6a0074d2accb1