request #10628: Detection and transformation to <a> tag of URLs in text can lead to XSS
This contribution also removes the linkification in text of the specific
subdomain www as it is very specific to the www subdomain and we can not
determine the proper URI scheme to use. Always using the http:// scheme
is not really future-proof as httpsis becoming the default.
This means if we find in a text "www.example.com" is will not be transformed
anymore to '<a href="http://www.example.com">http://www.example.com</a>'.
If the URL scheme is specified the transformation still happen, e.g.
a text with "https://www.example.com" is still transformed to
'<a href="https://www.example.com">https://www.example.com</a>'.

The target _blank as also be removed from the generated links as it is
considered a bad practice to modify the default browser behavior expected
by the user.

HTMLPurifier does not know the SSH scheme so to be able to accept
SSH URIs we need to define the SSH scheme in a way HTMLPurifier
is capable to deal with and to register the scheme we have defined.

Change-Id: I7424f672979d50e492b605be0f9a70c5da45b9f3