Display labels in homepage of pull requests
This is part of story #10498: Add labels on pull requests

Add a label to a pull request, the label should appear on the homepage of pull requests. You cannot edit labels from this page.

If you have a long title and many labels, then the labels should wrap before the title.

Quick access buttons are removed since they are not that useful.

Change-Id: I303bd55c3fd4d331cefece5f7d3f0a97fcd61239