story #10640 Have all mediawiki tables in one database
Introduce a new parameter in mediawiki configuration to allow usage of a
central database to group all mediawiki instances. The split is then done
using a table prefix.

ATM it doesn't change the way mediawiki are created (in a dedicated DB), you
will need to trigger a system event to kick the conversion:

    /usr/share/tuleap/src/utils/php-launcher.sh \
      /usr/share/tuleap/plugins/mediawiki/move_to_single_db.php \
      project_id

After that, everything should continue to work transparently (once the system
event is consumed). Especially the management of users and groups and site
admin stats.

Before moving to a central db, you will have to create a new db in your mysql
instance and granting your db users the approriate rights on it (see config
for more insights on how to do that).

Change-Id: Ied2d83421925cffafa3dc8116f3d756aeea27e1d