Handle anonymous users
This is part of story #1049: have base cross tracker search widget

Given a cross tracker search widget on a project dashboard,
When you navigate to the project dashboard as an anonymous user
Then you should see the report's selected trackers and the report artifacts
withtout error.
You should only see the selected trackers that are visible to anonymous users.
You should NOT be able to access the edit mode.

Change-Id: I356693dcdf26b05a626b004dd68d59256ea211ef