request #10652: Kanban columns are too narrow
If you have many columns (so much that there is an horizontal scroll in
the viewport) then the cards are barely readable, and long column names
are truncated without possibility to know the full name.

The proposal is to have a bigger minimal width of the column and to add
the attribute title on the column.

Change-Id: I15a9fd4c7eccfe974863164020bffb769349f370