request #10665: Reference published to a Bugzilla ticket can be duplicated
When a reference has already been published to a Bugzilla ticket it should
not be added anymore to this ticket.

For example if you reference a bugzilla reference in a commit message, push it,
then merge this commit into another branch and push it you should only have one
follow up.

Change-Id: I6f3a8e44b226656e2a39049bd25f9fc0b0c77b9c