Be able to display existing links in update release form
this is part of story #9692 Have a field in the FRS where I can add a link

In edit release form, we should be able to see existing links.
It's not possible to edit or delete them.

Edition will never be allowed, deletion will be done in a dedicated commit.

Change-Id: Ia09ad51ef040a908571d2458fdd11c2e96b9ff31