Add project label
Part of story #10501: Manage project labels

Go to project » admin » labels. You can add a new label with its own
style and color. You cannot add a new label with a name that is already
taken in the project.

There is no change in edition rules (labels can be merged).

Change-Id: I9dbac3206a5d09445d44c3528a1e78a69133c3d1