Refactoring: Clarify contract of labelable elements management
Part of story #10501: Manage project labels

Instead of having multiple dedicated events, we can clarify the contract of
LabelableDao and return the dao instead. That way, when we manage project
labels, we are more confident about the behavior of plugins: instead of
trusting the behavior behind three separate events (get, delete, merge) this is
now the responsibility of the core to execute respective queries, knowing that
the plugins respect the LabelableDao.

Change-Id: I556907cbeebb030e66bad8dd756fe91b35b58672