Encrypted Bugzilla API keys can be used
Existing cleartext keys are marked as such in DB
and Tuleap try to remove the cleartext version
when the configuration of a Bugzilla reference is
edited.

A future contribution will add a warning in the web UI for the
keys that has been stored in cleartext to encourage administrators
to rotate these credentials.

This is part of story #10606: encrypt bugzilla credentials (API key) before storing in the DB

Change-Id: I26208666f2ad3961021e3c12d4ba9926c3d6ae44