request #10650: Pullrequest list must not load all the pullrequest details
When we list the pullrequests for a repository in the pullrequest list
page, the full pullrequest representation is loaded. If you have a lot
of pullrequests and if the git repository is using fine-grained
permissions, the requests to load the pullrequests takes too much time.

Now we return only a minimal representation in git/:id/pull_requests so
that required information is still displayed on the pullrequest list
page (labels, author, title, …)

Change-Id: I2cbaf6217307bda6c1a787467c45962f63c3efc1