No need of nss & nscd for API tests
NSS management was introduced in API test suites to avoid errors related to
user and project management. With this commit, the problem is ditched entirely
by skipping the user and project management when the corresponding base
directories are not present on file system.

Incidentally this could be used to deploy platforms without NSS at all. It
might need some other developments.

Part of request #10713 REST test should not depend on Tuleap internal API

Change-Id: I107b8c3061c1bb9e45be209dff0ea386e79018d0