Make the subject customisation feature visible in the tracker email notification settings
This contribution only makes the feature visible. A future contribution
will add the possibility for a tracker admin to edit this
setting.

For now to enable/disable the feature you need to add/remove an entry
in the plugin_tracker_notification_assigned_to table.

This is part of story #10299: include [assigned to me] in artifact notification email

Change-Id: I0332c3f88ec005d4b603268c4b06e5b67ae7396e