Enabling [Assigned to me] in notification emails is done by tracker instead of by project
To enable/disable the feature for a tracker you need to add/remove
an entry in the plugin_tracker_notification_assigned_to table.

If you add previously enabled the feature for a project, all the trackers
of this project will have the feature enabled.

This is part of story #10299: include [assigned to me] in artifact notification email

Change-Id: I248d562520f7a978f0258c56249d7469d97d4069