Align widget icons top
This is part of story #10499: Have a labels widget on project dashboards

When your widget have so much labels configured that it must display
them on at least two lines in the title, then the widget icons (cog,
expand/collapse) should still be on the top of the widget.

This should work on IE11 as well.

Change-Id: I434d3a2b2c43d962fd292837e1ee01f1ad9ebbe6