Give site admin more control on mediawiki switch to version 1.23
Site admins can create a /var/tmp/tuleap_cache/STOP_SYSTEM_EVENT file to stop
execution if they need to stop it (because it's too long and block the main
system events queue).

Fix the way global switcher compute how projects are migrated to 1.23. Rely
on version registered in DB after switch rather than the allowed projects in
resource restrictor because the later will not be accurate if there is a failure
or if admins decide to stop the migration (previous §).

Part of request #10685 Migrate all mediawiki bases to 1.23

Change-Id: I63436fcd6d9d3e4f7e4e79f75bde8635acebdff0