Have linked bugs at initial load
This is part of myenalean #42386: see bugs I created / linked to a
execution failure

Depends on https://gerrit.tuleap.net/#/c/9859/

How to test:

Given a test execution linked to a bug artifact (and given you had
configured the bug tracker in administration), when you navigate to that
test execution in TTM, the list of linked bugs will be shown in the "bug
dropdown".

When you create a new bug or link to an existing bug, the change is not
taken in account, you need to refresh. Will be handled in a dedicated
commit.

Change-Id: I5dfdcf227b67e8087dccd6b37191d546d11f028a