Introduce @comment in TQL grammar
Part of story #9882: TQL in follow-up comments

No functional change in the Web UI.

When you parse a query like:

  field1_name = "f1" AND @comment = ""

Then you should have a Field with name = 'f1' and  Metadata
with name = 'comment' in the tree.  (see example by running
test_parser.php in grammar directory).

If you try to enter @comment in web ui, you will get an error.
If you try to enter @summary in web ui, it will use the standard field
summary (like if you entered 'summary = …'); will be fixed later to not
clutter to much the review.

Change-Id: I329a7635f9a6f367470d7e07bc5f4fa9bf642cac