Remove siteadmin Artifact link types restrictor
This is part of story #10736 administrate artifact links type at project level

There is no functionnal changes for the end user.
If the option is activated or not in the tracker global admin,
the types will appear/disappear in:

* Tracker report
* Artifact view (read-only and edit mode)
* CSV export
* XML import / export

Technically, all the check / calls are now done with the new objects.
The old database is no more instanciated, but the existing one are not
removed for now.

Change-Id: I397135a5a09ce19b1169a5d0f55a56dd7c7310f9