Refactoring: Rename InvalidFields to InvalidSearchables
Part of story #9882: TQL in follow-up comments

When we parse the query, we collect not only invalid fiels, but also
invalid metadata. Therefore rename fields to metadata during this
collection.

For the end users, we still display "The field xxx does not exist"
because its the closest term available to its understanding (we should
not talk about "searchable" or "metadata" in the ui).

+ Remove unneeded dependency on field factory
+ Enforce InvalidFields\* usage in InvalidSearchableCollectorVisitor to
  better highlights that the ComparisonVisitor only target fields.

No functional changes:
If we use @comment = 'toto', then an error is raised saying that
@comment is not supported.

If we use @summary = 'toto', then an error is raised saying that
@summary is unknown.

If we use summary = 'toto', then the query is performed successfully.

Change-Id: I261a7a97998c9e8a8b9fa3913ded88321ebcaa12