Rename getField in getSearchable
Part of story #9882: TQL in follow-up comments

No functional changes:

If we use @comment = 'toto',
then an error is raised saying that @comment is not supported.

If we use @summary = 'toto',
then an error is raised saying that @summary is unknown.

If we use summary = 'toto',
then the query is performed successfully.

Change-Id: I031881b342018a8c2c6da8101661c566539b8172