Enable and disable types per project
This is part of story #10736 administrate artifact links type at project level

All custom types could be enabled/disabled. Only the system type _is_child could
be enabled/disabled regarding the hierarchy (could be disabled only if no
hierarchy set in project)

For now, others system types provided by other plugins are not updatable,
and the changes done in this page are not taken into account in the tracker
service.

Change-Id: I6f52c7faed6425338088cd71b5565392c32258d0