setup.sh now install and configure in PHP 5.6 by default
+ manage upgrade from apache/5.3 to nginx/5.6 for tuleap-aio image

Note: tuleap-aio was keeping passwd, groups and shadow as it was done
at the origin of the docker image. However, this is not sustainable on the
long term as ids order are not guaranted. So switch to "fix-owners" stategy
from tuleap-aio-dev and tuleap enterprise.

Part of request #9202 Make Tuleap compatible with PHP 5.6

Change-Id: I02bc57d7835b70015f13e09211e12e60a5fbb116