Check type usage during changeset creation
This is part of story #10736 administrate artifact links type at project level

The type is now checked when the changeset is created.
If the type used is no more usable while editing / creating, an error
message is displayed (twice I know, but a next patch will fix that).

If the type previously used is no more available, the type will be clear
automatically. This behaviour is set in the warning content of the
administration page.

Change-Id: If9204ede8bc4c4a395a0edc4e4c23eb99ac6c496