Tracker table report take into account disabled types
This is part of story #10736 administrate artifact links type at project level

Now, tracker reports table allow to select only usable types as a
column.

If a previous column was set in the report, the option is still here to
remove it. Once removed and the report saved, it's no more possible
to select it.

Change-Id: Ia2249d9dd01a932c9f9dc05ce779f3600994559d