Import XML can deal with disabled types
This is part of story #10736 administrate artifact links type at project level

The import XML has the same behaviour as in the UI.

All types are activated by default, only the ones with is_used="0"
are disabled. The same goes for use-natures in trackers tag.

The import keeps the history of types provided in the file and disable
after artifact imports the types.

If _is_child is disabled, the hierarchy is not imported.

Change-Id: I024888f6cecd40543a4811ccac33870a9e2ebc48