Fix current pane for references pane
Part of story #10061 new project admin homepage and navigation

When you go to project » admin » references and you try to add or edit
a reference, then the tab "References" should be active.

We should not rely on a url parameter to know if a pane is selected or
not. As soon as we enter in the labels controller or the references
controller, we know that we are in the corresponding tab. Using an url
parameter is bad because it induces duplication and is error prone.

This commit applies this approach only for burning parrot pages, and for
references administration. Changing the pattern for every flaming parrot
pages would clutter too much the review, therefore we must let a default
parameter in project_admin_header() until we fix other pages :/

Change-Id: I55c7a2b8720108d92466eaaf8dac4d8aadf79725