Do not center permission request form
Part of story #10061 new project admin homepage and navigation

Go to project » admin » permissions » permission request

Centering the form was probably a good idea when the information were
enclosed in the legacy project admin homepage. It is not anymore the
case.

This commit also move the docman notice as an alert, and locate it
where it is relevant.

Change-Id: Ieeb8e314cbbfd70246f51aaa6b5ef46cde9881a6