Enable ES2017 and update karma configuration
This is part of myenalean #42386: see bugs I created / linked to a
execution failure

Gulp now only deals with angular-gettext strings extraction.
Upgraded karma-related packages.
Babel has two configs: one for IE11 and one for Chrome/Chromium (for
unit tests). The one for IE11 is used for the production build, the
other only for unit tests.

How to test:
- "npm run test" should pass
- "npm run coverage" should pass and show coverage
- "npm run build" should transpile ES2017 code for IE11
- "npm run watch" should pass and transpile ES2017 code for IE11

Change-Id: I6668911f00a2fea889a924e0853642e4af430913