See linked requirement in test execution
This is part of activity #42389 see link of the requirement linked to test

If an artifact that is not a test execution is linked to your test definition
and the typed link is _covered_by, this information is now displayed in the execution
view in a dedicated section.

If there is multiple links that match this condition,
we display the first one.

Change-Id: I7e7b2077333c3e274f2a9db716e29bf578bb3c8f