Realtime refresh linked bugs
This is part of myenalean #42386: see bugs I created / linked to a
execution failure

Depends on https://gerrit.tuleap.net/#/c/9859/
Depends on https://gerrit.tuleap.net/#/c/6499/ to deal with
artifact-only permissions

Also adds margin between the tlp-badge and the artifact title (also in
the linking modal).

How to test:
- Using 2 browsers/users, link a Test execution to a Bug artifact with
Alice. Bob should see the update in "Linked bugs" dropdown in realtime.
- Given Bob was not viewing the same execution as Alice, when Bob goes
to the same execution as Alice, the "Linked bugs" dropdown is up to
date.

Change-Id: I01cce34a5c9e16a6fa5d334ac5f001d1ec168a2c