Migrate tracker list in reading mode to Vue
This is part of request #10799 Migrate cross tracker widget in vuejs
Also needed for read-only query in story #10678 define a query for title
and description semantics

No functional changes.

The list of selected trackers in the report in the widget's reading mode
is migrated to vue. It should update correctly when:
- the widget is loaded
- you add / remove trackers in the writing mode
- you change the writing mode and hit cancel

Change-Id: I2570b0372d653e769a60a389652d263be3f5516d