GET frs_release/:id/frs_files
Part of request #10039: Retrieve/update FRS packages via REST

Information is already available through frs_release/:id, but this is a
first step to be able to add files to a release.

Please note the usage of array_slice for pagination: I know this is
generally bad to rely on array_slice for performance reasons. However
the average numer of files in a release is quite low. Furthermore all
files are already returned in frs_release/:id. Therefore, instead of
updating the internal api of FRS to have nice paginated resources, I
prefer to spend time in implementing missing gaps in REST API instead of
doing premature optimisation.

Change-Id: I7ee7875b50af1ee61a36153dd7b6a54c34b4114f