Display ugroup permissions in modal
Part of story #10062: Users & groups management overhaul

Go to project » admin » groups » LeGroup.

There should be a button to display permissions of the ugroup in a
modal. Actually this is the content (and the code) of the old pane
"permissions" which does not exist anymore.

No effort is made to make the retrieval of permissions cleaner as this
code will be removed when we will work for ugroup permissions in a next
story.

Change-Id: I86390b3297a708bd330bd8114190f7ca10ea0cd6