Introduce pane to see user group binding
Part of story #10062: Users & groups management overhaul

Go to project » admin » groups » LeGroup.

You should see a pane that display information about binding of a user
group, like it was done before in the dedicated pane.

Confirmation modal (with csrf) for binding removal will be done in a
dedicated commit.
Modals for ugroup binding (& ldap) will be done in a dedicated commit.

Change-Id: I24d0c66c45782f3e734efcb4f765c18f9ebb80cc