Add binding in a modal
Part of story #10062: Users & groups management overhaul

Go to project » admin » groups » LeGroup.

The button to add a binding now triggers a modal.

The code to handle the action on form submission as been copied from
Project_Admin_UGroup_UGroupController_Binding to
Project_Admin_UGroup_UGroupController and will be cleaned in a next
commit when we will have more visibility on refactoring scope.

Change-Id: I5923507108de053c8fe5c18b93a8bfe762bb3e38