Import members modal
This is part of story #10062 Users & groups management overhaul

This is the first part of the work on Import members modal. The goal
of this commit is to add a modal allowing the user to upload a file
containing users to import. The validation part of the workflow
will be moved in a BP modal in a next commit.

Change-Id: I6bd7e31a157268701ea3e21e001f3f8d6b3b0298