Manage ldap binding in a modal
Part of story #10062: Users & groups management overhaul

Now the binding of a ugroup to a LDAP group is done in a modal.

The old "Binding" tab has been kept in this review so that you can
compare the old and new behavior that should be really similar. This old
tab will be removed in a dedicated commit.

PS: Sorry for the big review :(

Change-Id: I4fcfbd703d6b6ad0932058713cc08e65e2da15f6