Use the tracker reference representation in the cross tracker search widget
Using the full tracker representation might cause performance issues with
trackers with a complex structure. More data than needed are retrieved which
could lead to slow download and parse of the large JSON retrieved or you can
even reach the memory limit server side while contructing the response.

There are possible optimisations to improve the response time of the server,
improving the memory usage seems to be more difficult. OPCache configuration
can be tweaked to some extends but the limit is still quite easily reachable
on instances with trackers using multiple large binded list fields.

This contribution solves the issue for the cross tracker widget by using a tracker
representation with only the needed information. This way the REST route is fast
and has a low memory usage.

request #10838: Cross tracker search plugin throw fatal errors when many users are in project

Change-Id: Ibbec9548f9ae38925f80ee83b505b01c09746674