request #10855 Translation files processing might lead to a php core dump
It seems that recent php version might lead to a core dump when there is
heavy load + writing and including php: https://bugs.php.net/bug.php?id=52752

This rewrite the way langs are stored to avoid a include. This comes with
a memory penalty of +2.5MB per request and +2.8ms.

Change-Id: I81894ac7f9bbe2d44541ceea4efb53e28ac22871