request #10865 Unable to comment an artifact that is using a hidden value of a required field
Given that you have a tracker with a required selectbox with the following values: [A, B, C]
And an existing artifact has the field selected to A
When in field usage you change the value A to "hidden"
Then you cannot update anymore the artifact unless you edit the field to choose B or C.

You should be able to update the artifact without having to change the value of the field.

The selected solution is to keep the hidden value in the form until its
removed by an artifact update. This value is obviously no more
selectable if not previously selected.

Change-Id: I4ecd98492c8ce1d66e0837e48a32bb6f71d86f30