Cache singular gettext lookup when called from a Mustache template
It reduces the wall time between 10% and 50% (depending of the number
of users) to load the project members page.

This contribution does not deal with lookup of plural strings as it is
a bit more tricky and not needed for the concerned pages.

This is part of story #10062: Users & groups management overhaul

Change-Id: I2477361765cba8b77c6d47d489fdb1d3f3daadc4