Improve performance of inline filter
This is part of story #10062: Users & groups management overhaul

How to test:
- Filter a table with many entries (several hundreds). The filtering
should be faster than before this patch.

To achieve this, I removed the ability to keep the scroll up when using
the filter (see 1f2e675cbfeb8a7f26adc32e26833af03b5d34c4) as it didn't
seem to work and was the major source of slowness.

Change-Id: Ide102e29d7b5b89d27957dc62e68a6de11ba134e