Only one modal to remove a user from a user group
This is part of story #10062: Users & groups management overhaul

Go to a user group and remove a user, a modal should appear and this is
the right user that is removed.

This will help when there are many users in the group.

Also added the ability to destroy (remove event listeners) a TLP modal.
Applied this to project members too so that it removes the DOMException
errors.

Change-Id: I436091fd83aefe65d799028b68438cc2b58794cc