request #10876: Query consistency is not checked before execution when creating a new cross tracker report
You can see the issue when you try to add a tracker where you do not have
all the needed rights to see the columns for example.

Since the cross trackers search widget is going to be fully dynamic
in future contributions, it does not feel necessary to try to filter
the projects/trackers that could be used in the dropdowns since
this filtering will soon not be possible.

Change-Id: I77d23acd0a72e65111f2aa763aeab9291ef95083