Burnification of project » admin » services
Part of story #10913: Services management

Only list of services of a project is displayed with BurningParrot. The
modals to create or edit a service will be done in dedicated commits.

There is no functional changes with previous behaviour except for a
column that has been removed because it was returning useless/inaccurate
data (conf inherited for project 100).

Change-Id: I45e8b7fd08fed4a9faddf9b1b7f1342e4a655672