request #10916: LDAP user with an UID matching multiples users according to sys_ldap_search_user param gets an error when they commit
The user is now retrieved only according to the LDAP UID as it is currently done
in the Tuleap SVN core integration (see src/utils/svn/commit-email.pl and
src/utils/ldap.pl:L109).

Also since this error is easier to encounter when a user with an LDAP UID
having spaces in it, the command given in the checkout helper has been updated
to take it into account.

Change-Id: I02a501b6bad579755466544253a5acf584042902