Be able to compute values
this is part of story #10670 a BurnUp chart based on Done stories

This commit introduce computation for Burnup chart.
In order to ease review, we didn't made the whole computation in one commit.
This commit only retrieve the  linked artifact (without _is_child link) with a Done semantic.
There is no sum of initial effort semantic for now, it will be done in a dedicated commit.

How to test:
create a hierarchy with some artifacts,
some with "Done" status, some with other status.

Access to an artifact with a burnup, the field is not displayed
You should see the list of all your open artifacts in /var/log/tuleap/burnup_syslog

Change-Id: If2a0fba7c1a6de5998bb9167132c45e595839c39