request #10923: Do not retrieve unneeded permissions
We don't display "permissions" button in dynamic ugroups administration
(Project administrators, …). However the code still fetch permissions
for those ugroups which is therefore useless.

Furthermore as fetching permissions has always be done on static
ugroups, the retrieval of permissions for dynamic ugroups is wrong: for
example it fetches permissions for **all** project administrators
ugroups instead of only for project administrators ugroup of the current
project. This produces a severe downgrade of performance (or fatal
errors) on big instances where there are lot of projects and lot of
permissions.

Change-Id: I6005ee75047d3bf212675f9900ff1816b4ec2281