Be able to compute values: add initial effort sum
this is part of story #10670 a BurnUp chart based on Done stories

Burnup field now display the value of burnup sum of initial effort for today.

How to test:
Given an agiledashboard with Release > Sprint > user stories
User stories must have Done (field status) and initial effort (field initial_effort) semantic set.
User stories field for initial_effort must be an integer field
In release and in Sprint initial_effort field must be computed.

We should be able to compute the sum of initial effort for every artifact with a Done semantic.

Change-Id: I94b9fd11df58c715862bca52d0c01b75f54b2f4c