request #10928: Make forking a Git repository more reliable
When forking a repository Tuleap should not expect the repo to be
perfectly configured.
Besides, the configuration should be handled by gitolite entirely,
not sometimes from the existing repo and sometimes from gitolite.

This is meant to avoid issue when forking a repository that has been created under
gitolite2 who still has links to gitolite2 hooks.

Change-Id: I41853256e04a93a62af662b0c0d0d4799c0366c9