Save query xtracker report (update REST API)
This is part of story #10678 define a query for title and description semantics

Using an empty query is allowed.
All the query checks are done before updating the report and
HTTP 400 are thrown when the query is invalid.

More REST tests will be added once the feature will be fully developped.

Change-Id: I40a4fe90319c8209b0f57db63f6c82c5f96d8ecc