Refactoring: delegate collection of errrors
Part of story #10678 define a query for title and description semantics

No functional changes in validation of the expert query in both trackers
and cross trackers.

In a next commit, we will validate that if a comparison is made against
@title or @description, then all trackers involved in the query are
using the corresponding semantic. This means that we will need to pass
the list of trackers id from the request to the collectErrors in the
cross-trackers case.

In order to produce small reviews, this commit only extract the
collection of errors in a dedicated object.

Change-Id: I5d5a034b6945b322dd60364dfe437ecca56b42d1