Reject query when one tracker does not use a semantic
Part of story #10678 define a query for title and description semantics

If you use @title (resp. @description) in your query and one of the
trackers does not have the title (resp. description) semantic, then an
error message is displayed to the end user.

If you use multiple @title (or @description) in the query, then the
error message is not duplicated.

Change-Id: I8d3e2c4ca8528eeda4cda6ff25e9d5307a9bb1aa