Sorting trackers on fields does not generate DB errors
It is no more allowed to sort reports using unsortable fields.

It is possible again to sort a report according to a priority field.

This solves request #10935: Sorting on "Rank" field causes artifacts
not to be displayed in tracker report.

The issue has been made more visible with the modifications made in
commits 2c50e75d8676fb44001712e446918f86d37c976a and
13308818555ef36fc832ad3989323ebbf28073db.

If an existing report has been sorted against an unsortable column,
sort will be ignored and a warning added instead of sort icon.

Change-Id: Ide5c559a1aacc2650e556cb672b4d415ac1696cd