Create basic REST test
This is part of story #10678 define a query for title and description semantics

This commit adds a first and basic test that get all the artifacts that
have a title.

In addition, this test will avoid regressions about report save.

To have a test that works, we have to add a space before @title in the
expert_query attribute in the REST PUT parameter, du to an issue linked
to Guzzle and @ parameter considered as paths[1]

[1] https://stackoverflow.com/questions/16812713/guzzle-post-parameter-issue

Change-Id: I4a0a5e2255fc2d97251f988513914cadf9a09442