Export/import done semantic in XML
This is part of story #10670 a BurnUp chart based on Done stories

If the done semantic is set in the tracker XML content, then it's imported.

As explained directly in the SemanticDone::save method, we call it only in the XML
import context (we do not want to also call it while saving the form to not
add feedback during XML import process).

In addition, the SemanticDoneFactory does the maximum checks possible
before providing the done values, so we assume that there are OK. And
given that we are in the middle of the XML import, the semantic done is
not yet created, se we cannot check for closed value only in XML import
...

This particular point must be fixed (if possible) in a dedicated next commit.

Change-Id: Ia45c8a629622bc3c807383464c668a7005affffc