Introduce SemanticDoneValueChecker object
This is part of story #10670 a BurnUp chart based on Done stories

The check of the validity of the value is extracted in a dedicated
object. There is no functionnal changes.

This object deals with the validation in the Web UI. It will be used to
validate data from XML in a next commit.

Change-Id: Ib9abd64a80305e4632ed885d044ce930a7898939